refactor(decompose): split packages/libraries/graph-tools/bin/vt-debug-stress.ts into folder
