+++ packages/libraries/graph-tools/bin/vt-debug-stress/io.ts
@@ -1,0 +1,11 @@
+import fs from 'node:fs/promises'
+import path from 'node:path'
+
+export async function writeJson(filePath: string, value: unknown): Promise<void> {
+  await fs.mkdir(path.dirname(filePath), { recursive: true })
+  await fs.writeFile(filePath, `${JSON.stringify(value, null, 2)}\n`, 'utf8')
+}
+
+export async function readJson<T>(filePath: string): Promise<T> {
+  return JSON.parse(await fs.readFile(filePath, 'utf8')) as T
+}
